fix for validation
git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@10145 8b29bb6e-589a-435a-9a12-3ab907be3462

+++ components/aracne-java/src/org/geworkbench/components/aracne/data/AracneInput.java
@@ -170,7 +170,7 @@ public class AracneInput {
 		this.consensusThreshold = consensusThreshold;
 	}
 
-	public double getConsensusThreshold() {
+	public float getConsensusThreshold() {
 		return this.consensusThreshold;
 	}
